Print warnings & errors to stderr
Not sure whether this message in status.sh is really a warning, but
all other messages being printed directly before a 'return 1' are
sent to stderr. Also, I've just run into a case where it would be
handy to keep the 'return 1' output separate from the 'return 0'
case. See the next commit for that.