Bail if cd fails
No sense chugging along if SERVER_DIR does not exist/can't be our cwd.

This has the unfortunate side effect of preventing read-only commands
from working if the cd fails for whatever reason. In the future, we
should sprinkle cd logic throughout actions that need it (start,
explore, map, etc.), but for now it's better to be safe than sorry.