Ease the use of `clicraft cmd`
Strip the first slash (/) of "/command" when using `clicraft cmd /command`.
This should make it more intuitive to use in most cases.

One case where this could cause confusion is for commands that use
two leading slashes/ WorldEdit's //replace, for instance, would have
to be invoked with `clicraft cmd //replace`, not `clicraft cmd /replace`
as was the case previously.