Removed the double clone
Removed the clone to clicraft-build as it had already been cloned
to clicraft making one of them redundent. As the clone to clicraft only
cloned if needed (otherwise it would pull) and clicraft-build did a
fresh clone each time it was better to keep the clone to clicraft.

If any of that makes sense... or if anyone ever read these things :D