Fixes error message and exit status of start
fix #8

The exit status and error message are now based on the result of the
status command. A sleep of 1 second was added to give minecraft sometime
to decide if it want crash.

The exit status is based of the the status command rather then the exit
status of tmux as tmux should return successful even if the command it
is executing fails.