Uncomment options in example clicraft.conf
As suggested by Michael Daffin in issue #7, uncomment the options in
clicraft.conf and add a warning so that users don't accidentally
set garbage options due to uninitialized default values.

Due to this change, `clicraft edit` also had to be modified to check
for changes instead of non-comments before creating a new file from
a template.

This is (hopefully) a stop-gap solution until a proper one can be
found.