Move all example actions into src/etc/action.d
The example actions in the scripts/ directory were placed there because,
unlike the examples in src/etc/action.d, they wrapped third-party software
and required that software to run. As more and more examples get added, it's
becoming clear that this distinction isn't useful, and makes documentation
more difficult. So put them alongside the other example scripts and install
them in the same way.

Also, remove the --with-mcexplore, --with-c10t, and --with-mmat options to
./configure, as they're no longer useful.